mellanox_os/src/requirements.txt
cloudshell-shell-core>=2.3.0,<2.4.0
cloudshell-networking>=4.0.0,<4.1.0
cloudshell-automation-api
paramiko